+++ src/main/java/org/jboss/netty/channel/socket/nio/NioWorkerPool.java
@@ -15,6 +15,8 @@
  */
 package org.jboss.netty.channel.socket.nio;
 
+import org.jboss.netty.util.ThreadNameDeterminer;
+
 import java.util.concurrent.Executor;
 
 
@@ -25,13 +27,22 @@ import java.util.concurrent.Executor;
  */
 public class NioWorkerPool extends AbstractNioWorkerPool<NioWorker> {
 
+    private final ThreadNameDeterminer determiner;
+
     public NioWorkerPool(Executor workerExecutor, int workerCount) {
-        super(workerExecutor, workerCount);
+        this(workerExecutor, workerCount, null);
+    }
+
+
+    public NioWorkerPool(Executor workerExecutor, int workerCount, ThreadNameDeterminer determiner) {
+        super(workerExecutor, workerCount, false);
+        this.determiner = determiner;
+        init();
     }
 
     @Override
     protected NioWorker createWorker(Executor executor) {
-        return new NioWorker(executor);
+        return new NioWorker(executor, determiner);
     }
 
 }
